+++ src/builder/builder.js
@@ -3,6 +3,7 @@ const Parameter = require('./parameter');
 const axios = require('axios').default;
 const Utils = require('@openeo/js-commons/src/utils');
 const ProcessUtils = require("@openeo/js-commons/src/processUtils");
+const ProcessRegistry = require('@openeo/js-commons/src/processRegistry');
 
 const PROCESS_META = [
 	"id", "summary", "description", "categories", "parameters", "returns",
@@ -117,21 +118,17 @@ class Builder {
 	 * 
 	 * Each process passed to the constructor is made available as object method.
 	 * 
-	 * @param {Array.<Process>|Processes} processes - Either an array containing processes or an object compatible with `GET /processes` of the API.
+	 * @param {Array.<Process>|Processes|ProcessRegistry} processes - Either an array containing processes or an object compatible with `GET /processes` of the API.
 	 * @param {?Builder} parent - The parent builder, usually only used by the Builder itself.
 	 * @param {string} id - A unique identifier for the process.
 	 */
 	constructor(processes, parent = null, id = undefined) {
 		/**
-		 * List of all non-namespaced process specifications.
-		 * @type {Array.<Process>}
+		 * A unique identifier for the process.
+		 * @public
+		 * @type {string}
 		 */
-		this.processes = [];
-		/**
-		 * Namespaced process specifications. EXPERIMENTAL!
-		 * @type {object.<string, Array.<Process>>}
-		 */
-		this.namespacedProcesses = {};
+		this.id = id;
 		/**
 		 * The parent builder.
 		 * @type {?Builder}
@@ -154,27 +151,53 @@ class Builder {
 		this.parameters = undefined;
 
 		/**
-		 * A unique identifier for the process.
-		 * @public
-		 * @type {string}
+		 * List of all non-namespaced process specifications.
+		 * @type {ProcessRegistry}
 		 */
-		this.id = id;
+		this.processes = null;
 
-		if (Utils.isObject(processes) && Array.isArray(processes.processes)) {
-			processes = processes.processes;
+		// Create process registry if not available yet
+		if (processes instanceof ProcessRegistry) {
+			this.processes = processes;
 		}
-		if (Array.isArray(processes)) {
-			for(let process of processes) {
-				try {
-					this.addProcessSpec(process);
-				} catch (error) {
-					console.warn(error);
-				}
-			}
+		else if (Utils.isObject(processes) && Array.isArray(processes.processes)) {
+			this.processes = new ProcessRegistry(processes.processes);
+		}
+		else if (Array.isArray(processes)) {
+			this.processes = new ProcessRegistry(processes);
 		}
 		else {
 			throw new Error("Processes are invalid; must be array or object according to the API.");
 		}
+
+		// Create processes
+		this.processes.all().forEach(process => this.createFunction(process));
+	}
+
+	/**
+	 * Creates a callable function on the builder object for a process.
+	 * 
+	 * @param {Process} process
+	 * @throws {Error}
+	 */
+	createFunction(process) {
+		if (typeof this[process.id] !== 'undefined') {
+			throw new Error("Can't create function for process '" + process.id + "'. Already exists in Builder class.");
+		}
+
+		/**
+		 * Implicitly calls the process with the given name on the back-end by adding it to the process.
+		 * 
+		 * This is a shortcut for {@link Builder#process}.
+		 * 
+		 * @param {...*} args - The arguments for the process.
+		 * @returns {BuilderNode}
+		 * @see Builder#process
+		 */
+		this[process.id] = function(...args) {
+			// Don't use arrow functions, they don't support the arguments keyword.
+			return this.process(process.id, args);
+		};
 	}
 
 	/**
@@ -188,30 +211,15 @@ class Builder {
 		if (!Utils.isObject(process)) {
 			throw new Error("Process '" + process.id + "' must be an object.");
 		}
-		if (namespace && namespace !== 'backend' && namespace !== 'user') {
-			if (!this.namespacedProcesses[namespace]) {
-				this.namespacedProcesses[namespace] = [];
-			}
-			this.namespacedProcesses[namespace].push(process);
+
+		if (!namespace) {
+			namespace = 'backend';
 		}
-		else if (typeof this[process.id] === 'undefined') {
-			this.processes.push(process);
-			/**
-			 * Implicitly calls the process with the given name on the back-end by adding it to the process.
-			 * 
-			 * This is a shortcut for {@link Builder#process}.
-			 * 
-			 * @param {...*} args - The arguments for the process.
-			 * @returns {BuilderNode}
-			 * @see Builder#process
-			 */
-			this[process.id] = function(...args) {
-				// Don't use arrow functions, they don't support the arguments keyword.
-				return this.process(process.id, args);
-			};
-		}
-		else {
-			throw new Error("Can't create function for process '" + process.id + "'. Already exists in Builder class.");
+		this.processes.add(process, namespace);
+
+		// Create callable function for pre-defined processes
+		if (namespace === 'backend') {
+			this.createFunction(process);
 		}
 	}
 
@@ -293,22 +301,14 @@ class Builder {
 	}
 
 	/**
-	 * Returns the process specification for the given process identifier and namespace.
+	 * Returns the process specification for the given process identifier and namespace (or `null`).
 	 * 
 	 * @param {string} id - Process identifier
-	 * @param {?string} [namespace=null] - Namespace of the process (default to `null`, i.e. pre-defined processes). EXPERIMENTAL!
-	 * @returns {Process}
+	 * @param {?string} [namespace=null] - Namespace of the process (default to `null`, i.e. user or backend namespace). EXPERIMENTAL!
+	 * @returns {?Process}
 	 */
 	spec(id, namespace = null) {
-		if (namespace && namespace !== 'backend' && namespace !== 'user') {
-			if (!this.namespacedProcesses[namespace]) {
-				return;
-			}
-			return this.namespacedProcesses[namespace].find(process => process.id === id);
-		}
-		else {
-			return this.processes.find(process => process.id === id);
-		}
+		return this.processes.get(id, namespace);
 	}
 
 	/**
@@ -336,7 +336,7 @@ class Builder {
 	 * @returns {boolean}
 	 */
 	supports(processId, namespace = null) {
-		return Utils.isObject(this.spec(processId, namespace));
+		return Boolean(this.spec(processId, namespace));
 	}
 
 	/**
@@ -344,7 +344,7 @@ class Builder {
 	 * 
 	 * @param {string} processId - The id of the process to call. To access a namespaced process, use the `process@namespace` notation.
 	 * @param {object.<string, *>|Array} [args={}] - The arguments as key-value pairs or as array. For objects, they keys must be the parameter names and the values must be the arguments. For arrays, arguments must be specified in the same order as in the corresponding process.
-	 * @param {?string} [description=null] - An optional description for the process call.a
+	 * @param {?string} [description=null] - An optional description for the process call.
 	 * @returns {BuilderNode}
 	 */
 	process(processId, args = {}, description = null) {