+++ CHANGELOG.md
@@ -10,9 +10,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Experimental support for process namespaces, see [API#348](https://github.com/Open-EO/openeo-api/pull/348).
 
+### Changed
+
+- Internally, a new process registry is used to manage and cache processes.
+- `listProcesses` doesn't cache requests any longer.
+
 ### Fixed
 
 - Return a better error message if issues with reading batch job results occur in `Job.getResultsAsStac`
+- `getAll()` functions return only properties for values that are defined
 
 ## [2.0.1] - 2021-07-14